KT-75880 [FIR] Use correct FirSession to expand non-library type aliases
When the wrong session is used, there
are nested type abbreviations formed.

Additional details:

Typealiases from library dependencies do not depend on the current
module, so we they can safely be resolved with the sessions
corresponding to their libraries, which allows us to resolve them even
when the current module is being analyzed at low phases
(such as COMPILER_REQUIRED_ANNOTATIONS, whereas for typealiases from the
current module we'd have to require SUPER_TYPES).

On the other hand, expandTypeAliases == true means we're resolving a
type with the current module having already been resolved to
SUPER_TYPES, meaning we can safely use the local session.
Using the local session is better, as this way we have more information:
resolving typealiases in platform sessions allows to expand them to
platform types, but resolving them in the common session can only expand
them up to an expect class.

^KT-75880 Fixed