[K/N] Fixed a bug with the liveness analysis fallback
During coroutines suspension, some variables are saved into the closure' fields.
The variables that are needed to be saved/loaded are computed with a special liveness analysis.
This analysis is a complicated algorithm, so to circumvent potential bugs in it, there is a
fallback mechanism - just save/load all the visible variables. This commit fixes one problem with it.

 #KT-75926