[FIR] Fix exception on invalid destructuring declaration
The code like `"hello".val.toString().println` produces weird PSI:
`val.toString()` becomes a destructuring declaration.

As a result, we end up with a synthetic `FirBlock` in the FIR tree
that contains the destructured properties. Normally, this block
is discarded during the Raw FIR phase if the PSI is well-formed.

We need to mark it as a fake source element, otherwise
we get exceptions in LL FIR.

^KT-73719 Fixed