fixup! [Gradle] KT-74735 remove internal CompositeProjectComponentArtifactMetadata
Revert changes again... it doesn't look like there's a reasonable way of making this work in a way that doesn't either cause Gradle to report 'task uses output of task' problems, or trigger compilation.

It doesn't look like there's a workaround for correctly registering task inputs, and triggering compilation is suboptimal but not breaking. So let's just trigger compilation.